Support systems without mmap()
We handle most mmap() problems via wrappers that do malloc() (for
private writable mmaps) or malloc()-and-pread() (for private read-only
mmaps).

CTF archive creation uses writable shared mmaps; for now, we simply
prohibit creation or use of CTF archives on non-mmap-supporting systems.
This might in future be lifted to allow use, but not creation.