ctf-lib: support pread() short reads, and systems without pread()
The ctf-lib code (including the newly-added mmap() fallback code) uses
pread(), but assumes that short reads never happen (that all such things
are errors).  This is a terribly dangerous assumption: they can and do
happen, even on regular files, and can very well succeed if retried.

Wrap pread() in a wrapper that uses read() in the absence of pread() and
retries on short reads until either there is an error, or EOF, or all
the data is read.