subr: only mmap() large CTF files
For smaller ones, the benefit of being able to mprotect() the regions is
limited due to their small size and consequent relative unlikeliness of
being overwritten by random wild pointer banding: it's dominated by the
cost of fragmenting the heap, adding VMAs and the like.

So get ctf_data_{alloc,free,protect}() to malloc() for regions under a
page in size.