archives, open: archive-member opening with symbol and string tables
Before now, users of libctf had to know whether they were opening a
ctf_file_t or a ctf_archive, and call the right function, or call each
function in turn until one of them succeeded.

This is a silly restriction brought about purely by the fact that the
struct ctf_archive is a direct mirror of the on-disk file format, so we
cannot add any flexibility.

So introduce an intermediate layer, not on disk ("struct
ctf_archive_internal"), and make all ctf_archive_t-taking file-reading
functions take one of these instead (conveniently, ctf_arc_write() does
not use ctf_archive_t in its prototype at all).

Typedef this new internal type to ctf_archive_t to avoid breaking
callers, and arrange for it to contain either a ctf_archive or a
ctf_file_t: if it contains a ctf_archive, it also drags around a symtab,
strtab and everything else needed to let us open archives with symbol
and string tables. Wrap every archive entry point so that if a name of
_CTF_SECTION (or, equivalently, NULL) is passed in, we hand back the CTF
file, if there is one: so sections and files containing raw CTF data
look just like CTF archives containing a single member. (The only thing
that insists on being handed a real CTF archive is raw archive
iteration, which is an archiver-only entrypoint that is not useful for
non-archives.)

Rejig ctf_open(), ctf_fdopen(), ctf_bfdopen() et al (but not the
pure-buffer functions ctf_simple_open() and ctf_bufopen()) to hand back
a ctf_archive_t rather than a ctf_file_t. Use changes from this:

  ctf_file_t *fp = ctf_open (name, &err);
/* error handling */
/* use it */
  ctf_close (fp);

to this:

  ctf_archive_t *arc = ctf_open (name, NULL, &err);
/* error handling */
  ctf_file_t *fp = ctf_arc_open_by_name (arc, NULL, &err);
/* error handling */
/* use it */
  ctf_file_close (fp);
  ctf_arc_close (arc);

but now it works for CTF archives as well as CTF files, and the caller
need not know the difference.

While we're changing this interface anyway, pass a target name into
ctf_open() and ctf_bfdopen(), so that callers can specify the target
explicitly if need be.

Changed functions (no callers yet):
  ctf_bfdopen() -> returns ctf_archive_t, not ctf_file_t
  ctf_fdopen() -> as ctf_bfdopen(): takes new target arg (may be NULL)
  ctf_open() -> as ctf_fdopen()
  ctf_close() -> in libdtrace-ctf only, when CTF_CLOSE_NEW_API is set,
                 takes a ctf_archive_t: otherwise, takes a ctf_file_t;
		 in binutils, always takes a ctf_archive_t

New functions:
  ctf_file_close() -> does what ctf_close() used to, and closes a
                      ctf_file_t

(Note: in libdtrace-ctf, ctf_close() works when passed *either* a
ctf_file_t *or* a ctf_archive_t: you can use it to close either. This is
essential for backward compatibility, since existing callers pass in a
ctf_file_t derived from ctf_bufopen() and expect it to work. In
binutils, use ctf_file_close().)