bfd: Only stash the ctfsect in ctf_bfdopen(), not ctf_bfdopen_ctfsect()
We should not stash the ctfsect for later freeing in
ctf_bfdopen_ctfsect(), since that function receives the ctfsect from
somewhere else, which owns it and expects to be able to free it.

We should stash it one level further up, in ctf_bfdopen(), which did the
allocation in the first place.