Signedness-correctness fixes on 32-bit.
To make 32-bit libctf usage work with format v2, we have to promote the
ctf_id_t to an unsigned long from a long, since we do many comparisons
with the parent/child range, which is precisely the place where unsigned
long and long differ (in 32-bit signed arithmetic, all parent types are
negative!)

So flip that, and make corresponding adjustments.  In order for error
checking to keep working, we have to cast CTF_ERR explicitly to a
(ctf_id_t) (which is safe, because the maximum possible type value has a
hole in it to allow for -1).  However, this now means that comparisons
with CTF_ERR are prohibited unless you are actually checking against a
ctf_id_t.  All other comparisons are in error in any case: despite the
comment in <ctf-api.h>, most error checking in libctf for functions
returning int rather than ctf_id_t already did a comparison against
0, not a check against CTF_ERR.