Drop unused stuff from the ctf_sect_t.
This does not need to be a mirror of ELF sections any more, now that we
use bfd to read those.

So drop fields we don't use: cts_type, cts_flags and cts_offset.  (Leave
the actual field definitions in libdtrace-ctf, and drop only the use of
them, for ABI reasons.  In binutils, drop it entirely, which also
removes the only off64_t from the public interface.)