ctf-api: drop IN_LIBDTRACE_CTF: use !BFD_ONLY instead
We cannot use a positive conditional to exclude compilation of things
in the public interface that should not appear in the binutils version:
users of libdtrace-ctf will not define the necessary #defines, and
compilation will fail.

Use a dependency on !BFD_ONLY instead. Existing users will not define
this, so it will appear, and all such code is unideffed out on sync
to binutils, as we want.

This effectively reverts commit 9c564566c0a6b027ccf8be82249660d5c33913b3.