Deduplicate the string table.
ctf.h says "the CTF string table does not contain any duplicated
strings", but this has never been implemented: the CTF internal string
table contains numerous duplicates, wasting hundreds of kilobytes of
space even when gzipped (the Linux 5.1.7 shared CTF file wastes 310KiB
on duplicates; vmlinux.ctf wastes 66KiB).

Deduplication has been made difficult by the way the string table is
constructed: the length is accumulated into fp->ctf_dtvstrlen at dynamic
string addition time (long before the string table itself is built, at
ctf_update() time), while the string table itself is built piecemeal by
pointer walking in code intertwined with the rest of the work
ctf_update() is doing.

So fix this with a new bunch of strtab utility functions that maintain
an atom table stored in fp->ctf_str_atoms.  Each atom contains a unique
string and a list of "refs", which are uint32_t * pointers to locations
that need updating with the strtab offset when the strtab is finally
constructed.  ctf_update() now works by repeatedly calling
ctf_str_add_ref() to add a new ref to a string, possibly instantiating a
new atom if the string has never been seen before: there is some code
reordering to make sure that this only happens once the data has been
copied into its final resting place, so that the ref update hits the
right locations rather than turning into wild pointer writes.  After the
walk over vars and datatypes accumulating their strings into the atom
table, we call ctf_str_write_strtab() to convert the atom table into a
strtab and update all the refs with the real strtab offsets, then purge
all the refs so that we can freely realloc() the buffer to stick the new
string table on the end of it, now we know how big it's going to be.
(We use a new realloc() wrapper that prohibits reallocation if there are
any outstanding refs, for safety.)

There is an extra wrinkle: if we just emit the atoms table by traversing
it in hashtab order and emitting it, the gzipped size of the CTF file
goes *up*, because random emission order is not good for compression.
Figuring out a good ordering compression-wise waits for a later commit.
For now, we ensure no degradation by emitting it in the same order it
was inserted: the same order it used to be emitted in.  We do this by
putting the atoms table members in a list in insertion order, and
emitting the table by walking the list.

Saved space in the gzipped vmlinux.ctfa: 294136 bytes.
No change to generated CTF.