Unhandled type kinds in init_types() are a corruption sign
We must detect failure to identify a type kind in this loop and report
corruption, or an unhandled type kind will circle round to the top of
the loop on a more-or-less garbage location at the end of the ctf_type_t
and the start of some type's vlen.  Disaster and bus errors will
probably result.