endianness: don't leave the header partially unflipped
In ctf_bufopen(), we endian-flip the separate copy of the header we use,
and we endian-flip the rest of the file, but we leave the copy of the
header in the CTF buffer unflipped.  We never use this buffer, so this
might seem harmless -- but if we write the file out again, the header
will be written out unflipped, so we'll have a file with a header
announcing one endianness and the rest of the file in the opposite.
Reopening such a frankenfile will be hard.

Fix trivial.