endian: copy foreign-endian CTF data into place even when uncompressed
We have three cases in the initial processing of post-header CTF data:
decompression (we allocate a buffer and copy into it), native endianness
and uncompressed (we just point straight at the buffer), and foreign
endianness and uncompressed (we allocate a buffer, and, uh... never copy
anything into it).

The latter case is rather wrong.  Do the missing copy.