slices: make aligned again
We cannot use unaligned slice types (or any unaligned types), since if
we do the entire file becomes unaligned and every single access from
that point on must use memmove().

In a future file format revision we will do exactly that, giving us
significantly more freedom: but for now, boost the size of slices to
prevent them misaligning the file.