util: 32-bit fix for ctf_realloc() error messages
The debugging message printed when ctf_realloc() failed was not 32-bit
safe: it printed a uint64_t with %li.  Unfortunately it seems unlikely
that %llu is portable enough, so use %lu.  It cannot possibly overflow
on a 32-bit system (there is not enough address space for that many
refs to accumulate): on a 64-bit system long is usually long enough,
and if it isn't, this is only a debugging message, not printed by
default, in any case.