header: add object and function indexes
The new cth_objtidxoff and cth_funcidxoff sections encode what used to
be one single string table into a pair of arrays of strtab indexes,
each one mapping the names in the corresponding section (data object,
function info) to the strings which comprise their names.

This is used by the compiler to encode the identities of symbols in
those sections, and stripped out at final link time.

(As before: at present libctf is probably nonfunctional with this change
alone in it if these sections contain any content, since no code has
been added to cater for it. Adjustments to do that will come in the next
couple of days.)