open: header upgrading and byteswapping for the new sections
This introduces code to allow the headers to change form between file
format versions, and uses it to shuffle a hole in the headers large
enough to insert empty function and object index sections.  This also
lets us shuffle the parmax stuff out of ctf_bufopen, where it really
doesn't belong, into upgrade_types, where it fits much better.

It also fixes a pre-existing bug as old as the introduction of the
variable section in 2012: we were checking the wrong header field to
detect the end of the function info section, so we would have walked
off it and treated variable info sections as function info sections,
surely resulting in a crash.  (However, no CTF files have ever been
generated that populate both sections, so this bug was never noticed and
is harmless.)

This breaks compatibility with CTF files produced with format v3 but
after the last release: all released versions of libctf and
libdtrace-ctf are still compatible.