open: rationalize the treatment of the CTF header
The CTF header handling in ctf_open() is complex and arcane, involving a
header that is usually left prepended onto the CTF buffer except when it
is upgraded, when it is copied out into a local temporary and later
discarded: the header in the buffer remains the old version, and not
endian-flipped: sometimes that version is maintained but the temporary
is not (the handling of the CTF_F_COMPRESS flag, for instance).  The
writing code writes out the buffer header even though it claims to be of
a version newer than its actual layout, rendering later opens
impossible. Also, because the header is so hard to access, it is not
dumped via the ctf_dump() machinery, unlike everything else.  The CTF
file as a whole (including header) is sometimes malloced and sometimes
received from the caller: we have a special freeing function for the
ctf_base that determines which is which.  We track some offsets that
include the header size and some that do not, with no indication which
is which.

This is too ugly for words. Change things so that the header is
routinely stored in fp->ctf_header(), upgraded and endian-flipped if
needed, and squeezed out of the ctf_base whenever it is rewritten (for
endian-flipping, upgrading, or whatever): write it out explicitly, and
never write out the header found in the CTF buffer, even if it hasn't
been squeezed out.  Keep a separate pointer to the freeable CTF buffer,
just like we do for the other possibly-freeable entities like the
ctf_parent.  Arrange for all sizes and offsets to exclude the header
size, just like they do in the resulting CTF file.

Used by the next commit.