link: merge the variable sections
This requires us to know, at variable-addition time, which type in the
destination container corresponds to the type of the variable in the
source container. What's more, the variable may already be there,
pointing to the right type, in either the parent or the child: or it may
have a conflicting name in the parent and need a new definition in the
child.

We solve all of this via a new mapping table stored in destination
containers' fp->ctf_link_type_mapping, which maps source container
(fp pointer, type index) pairs to type indexes (not IDs: the values are
unaffected by whether this is a parent or child container, and need
adjustment via LCTF_INDEX_TO_TYPE() to turn them into a type ID).
This mapping table is updated by ctf_add_type() at every stage of its
recursive operation, and then used to determine what type variables
should reference.  (It will also be used for the function and symbol
index section work, later.)