headers: demote <sys/param.h> inclusion to ctf-impl.h
Including this in <ctf-api.h> is far too disruptive, since <sys/param.h>
is a hoary old horror of a header that defines things like MIN and MAX
that callers reasonably expect to be able to define themselves without
clashing.

No negative effect on any current users (and, frankly, anyone who
was depending on this transitive inclusion was buggy anyway).