hash, link: drop the hash size trickery and just hardwire what we need
It's less elegant, but this works, while the generic approach led to
endless pain.  I'm not willing to waste *that* many hours debugging
a trivial data structure like this, to save this few lines.

Just hardwire the hashing function we need. It's horrible but will
probably work well enough for our limited purposes.