iter: make ctf_type_iter_all public
This means anyone can use it in dumpers.  (It doesn't make non-root
types meaningless, because you still can't look them up by name: but you
can now iterate over them if you need to.)

The callback it takes is a bit different: it takes a nonroot flag so the
function called can tell if a type is root-visible or not, since this
property isn't visible through the normal API (because it's not
something that is usually useful to know).