open: loosen the restrictions on string and symbol tables
If you provide an external symbol table at ctf_bufopen time, you must
provide an external string table (because that's where the symbol names
are stored).  But the converse is not true: you do not have to provide
a symbol table if you provide a string table, because you might be using
the string table to look up strings in a CTF file that doesn't contain
any symbol references at all, or which contains some but you don't plan
to use them.