Add per-CU mapping machinery.
This lets you transform the names used for per-CU CTF dictionaries in
the output of links, so you can group the output in some fashion other
than by translation unit.  (This will obviously only work in conjunction
with some custom consumer that is following the same rules.)

Utterly untested so far.