link: make sure ctf_link_outputs exists when adding a CU mapping
ctf_link_add_cu_mapping adds to ctf_link_outputs, so we have to make
sure that hash is created first.

(This does mean that you can no longer add to the link after you call
ctf_link_add_cu_mapping, but since ctf_link_add_cu_mapping has one known
user so far and it works in that order, this is not a real limitation at
present. We can lift it easily if it causes trouble later on.)