link: apply the exponential explosion fix kludge to ld links too
It is quite possible to get exponential explosions linking ordinary
things with ld.  Doing periodic ctf_updates isn't going to break
anything for ld and might help, so let's do it.  (This will all go
away when the deduplicator is rewritten.)

Also do it whenever we exceed the limit, even if we are within one
single input file.  Insofar as this changes semantics, it will
fractionally improve the quality of ctf_add_type's terrible
deduplication, and it helps with massive modules like the kernel's
nouveau.ko, which has >200,000 types and we really do *not* want
an O(n^2) explosion with n = 200000.