create: close huge memory leak
ctf_update() calls ctf_bufopen() to open the new container with a
dynamically-allocated, uncompressed buffer.  Normally this would be kept
around in case the caller was using it, but ctf_update() tries to trick
ctf_bufopen() into closing it instead.  But after the commit titled "open:
rationalize the treatment of the CTF header", this trick no longer works
and there is a much more straightforward way to do the same thing.

Switch, saving huge amounts of memory on large many-ctf_update() things
like dwarf2ctf runs (or ctf_link runs with the current performance
kludge in place).