create, string: handle external strtabs in var sorting and name addition
The external strtab code added a few weeks ago had a rather serious bug.
At ctf_update time, we emit the new strtab using ctf_str_write_strtab,
which traverses all strings with references to them in the atoms table,
generates the internal strtab for all that are not external references,
and points the externally-referenced ones to the offsets we have been
told the external strtab will be using.  It's a shame that  ctf_update
then proceeds to sort the variable table using a function (ctf_sort_var)
which explicitly references the internal strtab by name!  We should
clearly be checking for an external strtab here rather than looking up
offsets in the internal strtab which aren't even meant to point there.
At best, this leads to incorrect sorts (and, since this table is
bsearched, inaccessible variables and unspecified behaviour at lookup
time).  At worst, we fall off the end of the internal strtab and
coredump.

But we can't just use the external strtab here.  It is not available
yet: not as a string table, anyway.  The linker literally does not have
it to give to us: all it has is a hashtab much like our atoms table,
which it converts directly into a strtab it writes straight to disk.
So, instead, when we write the strtabs out, update a new dynhash
ctf_syn_ext_strtab which maps offsets in the external strtab to strings
in the atoms table: ctf_strraw can then look them up from there if
available, and avoid accessing the nonexistent external strtab.
Because this dynhash only contains external strings that we actually
reference, it consumes minimal storage (it doesn't even need to
duplicate the strings, since atoms in the atoms table that are present
when a ctf_update is called never go away until ctf_file_close).  This
suffices to fix things for the name-interning code so that it can just
call ctf_strraw and things will always work: we no longer need to
special-case and ignore returns of ECTF_STRTAB but can propagate them to
the caller (which is good because they might be legitimate).

But that's still not quite enough!  When ctf_sort_var is called, the
internal strtab is not in the usual place: we haven't allocated the new
ctf_file_t to stick it in yet (this is why we never used ctf_strptr or
ctf_strraw in ctf_sort_var before now).  So we add a new function
ctf_strraw_explicit which takes an optional explicit internal strtab to
use in place of the normal one, reimplement ctf_strraw in terms of it,
and call it directly from ctf_sort_var.

With all this in place, we can finally link files containing variable
sections that contain external strtab references without either grossly
mis-sorting things or crashing.