link: rationalize error handling
Internally to ctf_link*(), we were trying to return errors by hand via
returning ECTF error codes, rather than using the ctf_errno machinery
and risking leaking internal errors out to the outside world.

Given that this is harmless (nobody should even consult the ctf_errno
unless we return -1, and if we do that we must have set it to a suitable
value beforehand), this was quixotic at best.  It turns out it also
doesn't work, because the error codes are *positive*, so you cannot test
for a direct error return of one of those codes by checking for a
negative return value.

Just use ctf_errno on the main output CTF file instead.  This
necessitates a little transfer of errors into the main output CTF file
from sub-files, but really not very much of it, and errors are now
properly propagated.