create: correctly override external strtab with synthetic one
The recent addition of synthetic strtabs was causing ctf_update to fail
(but the failure was covered up by the link-time error handling bugs).
The problem is that adding in the synthetic strtab after ctf_bufopen
time is too late: ctf_bufopen itself needs it (or, rather, init_types,
which it calls does), but ctf_bufopen is the thing which *creates* the
fp, so we can't just force it in.

Instead, add new internal-only variants of ctf_simple_open() and
ctf_bufopen() which take an optional synthetic external strtab, and
initialize it properly.  (ctf_simple_open() and ctf_bufopen() themselves
become trivial wrappers around these new functions.)