link: make sure that resolved type conflicts never propagate up
A caller that erroneously called ctf_errno() when it did not get a -1
from a called CTF function might have seen a spurious ECTF_CONFLICT
error from a successfully resolved type conflict.  Forcibly set it back
to 0 / no error to emphasize that this sort of thing is normal operation
and not an error at all.