link: redo variable addition
The variable addition code was perfectly written, *if* we were called on
child CUs of the output link and had to map inputs to it.  But we're
called on child CUs of the *input* link -- so the whole thing was
broken.

Rewrite it to work properly, attempting an addition to whichever of the
parent or child the type mapping suggests (child if the type mapping
says the type is also in the child, otherwise the parent).  If we get a
name clash adding to the parent, because another variable exists with
the same name but a different type, create a child if need be, transform
the type ID into a child-relative one, and try adding again.  (If we
still fail with a clash, this is almost certainly because of something
like a CU type mapping change jamming multiple CUs into one: we don't
want such things to be errors because the Linux kernel ctfarchive is
rife with them, so just silnetly skip them.)