create: handle late additions of forwards in ctf_add_type
ctf_add_type() allows you to add a forward and then add a struct, union,
etc, but the linking machinery also wants to be able to add a struct,
union etc and then add a forward (conceptually, this should be a NOP:
the forward should disappear and you should get the original type back).

The lookup of forwards in ctf_add_type is also broken: we look up the
name in the global namespace, which means that if you have a
non-struct/union/enum type named FOO you cannot add a forward to
a struct, union, or enum named FOO either, even though they are in quite
distinct namespaces and the forward even carries round an indication of
what namespace it belongs in.

When ctf_add_type is creating a forward, it also neglects to add that
in: so for forwards to everything but structs, the forward will be
created in the wrong namespace and will then fail to properly unify with
its union/enum later on.