link: get the parent name right
In some situations (notably, when CTF is linked into kernel modules)
DTrace depends on the parent name being set correctly.  We were leaving
it set to PARENT, which is always wrong.

Set it to .ctf by default, and change it in all children later on if the
member name changer alters the name of .ctf to something else.