link: prevent lingering non-errors from causing spurious link failures
Many "errors" observed deep in the linking process are handled and are
not errors any more: as usual in libctf, if the called function did not
return -1, there is no error.  But ctf_link_one_input_archive is called
via ctf_dynhash_iter, which cannot return an error code, so the only
error communication mechanism is the ctf_errno.

So make sure the ctf_errno is set to zero when there is no error, or
ctf_link will spuriously claim failure when actually nothing went wrong.