create: explicitly track additions underway by ctf_add_type
Our previous attempt to keep additions of self-referential structures
from recursing infinitely by simply handing back all types that
ctf_add_type had already added had a disastrous bug: because forwards
get added here as well, we could never add any members to a forward
because it would always get handed back straight away!

We can avoid that problem by verifying that the src and dst types are of
identical size, but while this serves to speed things up quite a lot it
also brings back the original problem, because when the problem strikes
we are still adding members to the structure so it has obviously not
reached its final size yet.

So add an explicit hash of things being processed, ctf_add_processing:
turn ctf_add_type into a wrapper that constructs this hash and empties
it out (we don't want to take the hit of *allocating* it every time),
and add every type being processed to this hash before we do the
kind-specific processing that involves calling ctf_add_type on it.  We
can then solve the recursion problem by simply immediately returning any
type ID if the type is in this hash, confident that a higher level of
recursion will finish the job for us. Old items linger until this
ctf_add_type is completely done, but that's not a problem, since they
*have* just been added and aren't going to change until ctf_add_type has
returned.