Dynamic type rejig, removing the expensive ctf_update
The method of operation of libctf when the container is writable has
before now been that you add types to the dynamic type section, a linked
list, a hash of IDs -> dynamic type definitions (and, recently a hash of
names): the DTDs are a bit of CTF representing the ctf_type_t and ad hoc
C structures representing the vlen.  Whenever you want to look up types
by name, or write the file out, or (until very recently) reference types
you added by ID at all, you call ctf_update, which serializes all the
DTDs into a CTF file and reopens itself. The ctf_updated types are then
frozen in amber and unchangeable: all lookups will return the types in
the static portion in preference to the dynamic portion, and we will
refuse to re-add things that already exist in the static portion (and,
of late, in the dynamic portion too).  The libctf machinery remembers
the boundary between static and dynamic types and looks in the right
portion for each type.  Lots of things still don't quite work with
dynamic types (e.g. getting their size), but enough works to do a bunch
of additions and then a ctf_update, most of the time.

Except it doesn't, because ctf_add_type finds it necessary to walk the
full dynamic type definition list looking for types with matching names,
so it gets slower and slower with every type you add: fixing this
requires calling ctf_update periodically.

This is clunky and very slow but kind of works, until you consider that
it is in fact possible and indeed necessary to modify one sort of type
after it has been added: forwards.  These are necessarily promoted to
structs, unions or enums, and when they do so *their type ID does not
change*. So all of a sudden we are changing types that already exist in
the static portion.  ctf_update gets massively confused by this and
allocates space enough for the forward (with no members), but then emits
the new dynamic type (with all the members) into it.  You get an
assertion failure after that, if you're lucky, or a coredump.

So this commit completes the job started a few months ago and rejigs
libctf into something simpler.  We now use the dynamic type definitions
exclusively in writable containers, and the static type definitions
exclusively in readable containers.  The ctf_dtbyname hash I introduced
a few months ago, which maps things like "struct foo" to DTDs, is
removed, replaced instead by a change of type of the four containers
which track names. Rather than just being (unresizable) ctf_hash_t's
populated only at ctf_bufopen time, they are now a ctf_names_t
structure, a pair of ctf_hash_t and ctf_dynhash_t, with the ctf_hash_t
portion being used in readonly containers, and the ctf_dynhash_t being
used in writable ones.  The decision as to which to use is centralized
in the new functions ctf_lookup_by_rawname (which takes a type kind) and
ctf_lookup_by_rawhash, which it calls (which takes a ctf_names_t *.)

This change lets us switch from using static to dynamic name hashes on
the fly across the entirety of libctf without complexifying anything: in
fact, because we now centralize the knowledge about how to map from type
kind to name hash, it actually simplifies things and lets us throw out
quite a lot of now-unnecessary complexity, from ctf_dtnyname (replaced
by the dynamic half of the name tables), through to ctf_dtnextid (now
that a container's static portion is never referenced if the container
is writable, we can just use ctf_typemax to indicate the maximum type:
dynamic or non-dynamic does not matter, and we no longer need to track
the boundary between the types).  You can now ctf_rollback() as far as
you like, even past a ctf_update or for that matter a full writeout; all
the iteration functions work just as well on writable as on read-only
containers; ctf_add_type no longer needs expensive duplicated code to
run over the dynamic types hunting for ones it might be interested in;
and the linker no longer needs a hack to call ctf_update so that calling
ctf_add_type is not impossibly expensive.

You can still call ctf_update if you like (it's public API, after all),
but its only effect now is to set the point to which ctf_discard rolls
back.

Obviously *something* still needs to serialize the CTF file before
writeout, and this job is done by ctf_serialize, which does everything
ctf_update used to except set the counter used by ctf_discard.  It is
automatically called by the various functions that do CTF writeout:
nobody else ever needs to call it.  Finally, the serialization machinery
in libctf is an internal implementation detail!  I've been wanting this
for *years*.

(And, though this is not yet tested, it should make it possible for
forwards to work, rather than crashing.)