link: save a little memory
The memory consumed by all the linker state during a large link is quite
painful.  We can save on it by noting that we are one-pass, and once we
transition away from an input file, we never go back to it again: so the
contents of the mapping tables become useless and can be discarded.