create: don't lose track of the currently-processing hash in ctf_add_type
The ctf_add_processing hash is stored on the source fp, and contains
type IDs (so you can store parent IDs in there too).  But the src_fp
we pass down to ctf_add_type recursive calls is adjusted if a ref to
a parent type is found, to point to the parent instead: this ensures
that we can never accidentally look in some child for types cited by a
parent.  But this means that as soon as we look for a parent type,
the ctf_add_processing hash goes to NULL!

So pass the fp we are using to track the currently-processing hash down
explicitly, and do not change it when we look in parents.  (Its contents
are unambiguous because they cite parent types by ID explicitly, so this
is safe.)