create: populate the ptrtab for dynamic dictionaries too
After commit 3f62125f164797, we do not do all of ctf_bufopen for dynamic
dictionaries: we skip init_types, type upgrading, and symtab and ptrtab
initialization.

This is mostly accurate (init_types relates to extracting things into
the static type table portion, which we do not care about for dynamic
containers any more: the symtab initialization portion relates to
symtabs which cannot be present for dynamic containers), but we *do*
need to initialize the ptrtab or ctf_lookup_by_name will fail for
pointer types in dynamic containers.  This was always done badly --
newly-added types didn't appear in the ptrtab until ctf_update was
called -- but now it's not being done at all.  Applications notice.
So fix this by keeping it up to date at all times, in keeping with
the "ctf_update delenda est" philosophy.

This is mostly very cheap and noninvasive, since we only need to update
it when a new pointer is added, and that's only done in one place; but
we do need to make sure that the array is large enough.

To avoid having to constantly realloc the ptrtab, we start it at 1024
types in length (small enough to not incur disastrous overhead if there
are many open dicts, large enough not to require any growth at all for
most dicts) and grow it by 25% at a time.