types: handle dynamic encodings the same as static ones
If the encoding of a dynamic type is requested, we were always returning
the encoding from the dtu_enc member.  This is unfortunate, because for
slices it's stored in the (overlapping) dtu_slice member of the union,
and for other types we should return -ECTF_NOTINTFP rather than
returning a pile of total garbage from the wrong part of the
discriminated union.

Fixes test/unittest/funcs/tst.index.d, among others.