string: make it possible to store all type names in ctf_type_t.ctt_name
Before now, since time immemorial, there have been two ways of storing
type names in libctf: in read-only containers (and that portion of
containers which had been ctf_updated), the name was stored in struct
ctf_type_t's ctt_name member; in dynamic type definitions, even though
they contain a ctf_type_t embedded within themselves, the ctt_name is 0
and the name is stored in the dtd_name member of the DTD itself.

This has massive negative consequences as soon as we ditch ctf_update
and hope for dynamic types and static types to work the same way.  Every
single place using type names in CTF needed to be bifurcated to check
the dtd's dtd_name member as well, and this was not something being done
in one central place: it happens everywhere.  This is not a safe or
maintainable thing to do.

Instead, exploit the atoms table and the idea behind the recently-added
synthetic external strtab to make dynamic and static types more similar:
both of them now store their names in ctf_type_t.ctt_name, and the
dtd_name member is gone. ctf_strptr works for all type names, static
and dynamic.

This would have been impractical in the past: ctt_name is a strtab
offset and the strtab was assembled piecemeal: then, after my revamp for
strtab dedup, it was assembled in ctf_update (now ctf_serialize), which
is long after the DTDs are put together.  But we can fix this by adding
a new "provisional strtab" hash, fp->ctf_prov_strtab, like the synthetic
strtab a hash of offset to a string in the atoms table.  (This also
removes one string duplicate, saving a bit of memory at writeout time.)
All non-external strtab strings are added to this, and a "provisional
offset" (just a unique number higher than the cts_len of the real
strtab) is maintained to keep track of the last provisional string
added.  ctf_strraw now checks this hash for all strings with offsets
greater than the strtab length and less than the provisional offset.

We also need to ensure that these new DTD ctt_name references are
updated to point to the real strtab entries when we write the strings
out (possibly "real" entries in the synthetic external strtab, if they
are deduplicated against the ELF strtab), since you are allowed to write
the CTF file out and then do more work and then write it out again, just
as in the past you could do ctf_update, then more work, then another
ctf_update.  We do this by adding the address of the ctt_name in the DTD
as another reference to each string: the existing ref update machinery
that updates the references to strings in the final ctf_serialize'd
buffer will then also update the reference to the ctt_name.  We do need
to add a bit of extra machinery to let us delete refs, since DTDs can be
freely deleted and we don't want to drag around the ref forever if we do
that.

We also adjust things so that internal strtab entries with no refs do
not get written out: this is to ensure that deleted DTDs don't get their
strings written out to the strtab if nothing else is using them.